Move categoryfilter to components directory
